requirements.txt
typer[all]
uvicorn
fastapi[all]
uvicorn
httpx
pydantic
python-dotenv
sqlalchemy 
python-dotenv
pymysql
cryptography